init commit
systemgmi serves the neofetch logo on index and links to a TODO/WIP unit
list page, as well as some dead links. Other than that, the roughly 500
lines of C are sound with no memory leaks that I can easily see.
Threading is not implemented as efficiently as it should be (and there
will be some race conditions with the current implementation), but the
source architecture is written and it's actually doing stuff.