handle URL's without a trailing slash
If a URL was passed that looks like `gemini://localhost`, the loop
finding without_host would never reach a trailing slash, and, as such,
it would segfault as it read beyond the bounds of the string. The fix is
to have the loop look for either a trailing slash *or* a NULL
terminator, both of which may indicate the end of the hostname.