cache running systemd units on startup
This is a large commit containing a number of changes. See the diff for
implementation, but the big things are as follows:

+ A dependency on systemd is introduced. This just uses sd-bus,
  systemd's implementation of the dbus protocol, not any systemd-
  specific functionality.
+ The dbus session bus is connected to at startup and disconnected to at
  shutdown.
+ All running units are read from dbus into (heap-allocated) memory and
  deallocated at exit.

For the last change, the units are just read into an array. That's going
to have sub-optimal performance characteristics once we need to index it
by name (linear string comparisons for each unit), but it's fine-enough
for now.

Note that this commit does not introduce any way to access the data, it
just stores it.