display units on route /systemctl/list-units/
This writes out a simple gemtext list of the names of units on
ROUTE_LIST_UNITS. This patch makes the struct sd_unit public (so as to
avail its internals to formatting functions) and introduces a struct
sd_unit_iterator, which encapsulates the collection used in the systemd
compilation unit's context struct.

The current implementation is safe with regard to memory, but leaves
open the possibilty of race conditions if any modification to the
systemd context is made (as will happen when connections are served in
parallel and systemd's signals are listened to).