define and fix behavior in com/util/com_until_crlf
The previous implementation of com_until_crlf had a variety of issues
--- it was slow and expected that clients formulate input properly. If a
client requested a resource (or sends nonsense) that doesn't end with a
CRLF, the implementation would propogate a segfault up to the rest of
the program without a necessarily clear origin. Part of fixing this is to
have clear behavior when bad input is recieved. Namely, returning NULL.

In addition, this makes com_until_crlf O(1) in the ideal case
(well-formulated client input) and O(n) in either degenerate cases
(extra garbage after the CRLF) or incorrect cases (spec-non-compliant
clients that don't end their requests with a CRLF).