serve requests in parallel
This commit implements a thread pool that serves requests in parallel.
The old strategy was to serve requests serially in a thread parallel to
that which launched it. That approach was honestly probably good enough,
but this is definitely the more 'technically correct' implementation.

A worker thread is spawned for each CPU (or 32 threads are spawned,
whichever is smaller), and work is queued round-robin to them by the
manager. The first thread spawned gets the first queued job, then the
second, then the third, etc. Because jobs are short-lived, this should
keep the pool balanced.